Fix string writing function in Money
