+++ modules/squants/src/test/scala/pureconfig/module/squants/SquantsConvertTest.scala
@@ -6,6 +6,7 @@ import com.typesafe.config.ConfigFactory
 import org.scalacheck.{ Arbitrary, Gen }
 import pureconfig.{ BaseSuite, ConfigConvert }
 import pureconfig.generic.auto._
+import pureconfig.module.squants.arbitrary._
 import pureconfig.syntax._
 import _root_.squants._
 import _root_.squants.market._
@@ -78,8 +79,9 @@ class SquantsConvertTest extends BaseSuite {
   checkDimension(time.Frequency)
   checkDimension(time.Time)
 
-  it should "parse Money" in forAll { (m: Money) =>
-    checkConfig(SquantConfig(m))
+  {
+    implicit val mc: MoneyContext = defaultMoneyContext
+    checkArbitrary[Money]
   }
 
   case class SquantConfig[T](value: T)
@@ -95,28 +97,5 @@ class SquantsConvertTest extends BaseSuite {
   def checkConfig[T](config: SquantConfig[T])(implicit cc: ConfigConvert[T]) = {
     val configString = s"""{value:"${config.value.toString}"}"""
     ConfigFactory.parseString(configString).to[SquantConfig[T]] shouldEqual Right(config)
-  }
-
-  def quantityAbitrary[T <: Quantity[T]](dim: Dimension[T]): Arbitrary[T] = {
-    Arbitrary(
-      for {
-        n <- Arbitrary.arbitrary[Double]
-        u <- Gen.oneOf(dim.units.toList)
-      } yield u(n))
-  }
-
-  // Money.units is not implemented so we need an explicit Arbitrary
-  implicit val moneyArbitrary: Arbitrary[Money] = {
-
-    // BTC is not included: fails on input: 0E-15
-    val currencies =
-      List(USD, ARS, AUD, BRL, CAD, CHF, CLP, CNY, CZK, DKK, EUR, GBP,
-        HKD, INR, JPY, KRW, MXN, MYR, NOK, NZD, RUB, SEK, XAG, XAU)
-
-    Arbitrary(
-      for {
-        n <- Arbitrary.arbitrary[Double]
-        c <- Gen.oneOf(currencies)
-      } yield Money(BigDecimal(n).setScale(c.formatDecimals, BigDecimal.RoundingMode.HALF_EVEN), c))
   }
 }
